Add setPerformanceHintEnabled(), Dynamic CPU Load test (#1788)
Add ADPF support to Oboe

OboeTester: test dynamically changing load
Add an Activity that varies the workload.
Also add CPU affinity UI.
Also try to measure the time it takes for the CPU frequency to recover.
Add Synth from SynthMark for more realistic workload.
Use dummy workload when not stereo.
Make workload an integer (#voices).

+++ apps/OboeTester/app/src/main/java/com/mobileer/oboetester/AudioStreamTester.java
@@ -33,7 +33,7 @@ class AudioStreamTester {
     }
 
     public void reset() {
-        setWorkload(0.0);
+        setWorkload(0);
         requestedConfiguration.reset(); // TODO consider making new ones
         actualConfiguration.reset();
     }
@@ -46,7 +46,7 @@ class AudioStreamTester {
         mCurrentAudioStream.startPlayback();
     }
 
-    public void setWorkload(double workload) {
+    public void setWorkload(int workload) {
         mCurrentAudioStream.setWorkload(workload);
     }
 }
