+++ apps/OboeTester/app/src/main/java/com/mobileer/oboetester/WorkloadView.java
@@ -97,10 +97,10 @@ public class WorkloadView extends LinearLayout {
     }
 
     private void setValueByPosition(int progress) {
-        double workload = mExponentialTaper.linearToExponential(
+        int workload = (int) mExponentialTaper.linearToExponential(
                 ((double)progress) / FADER_PROGRESS_MAX);
         mAudioStreamTester.setWorkload(workload);
-        mTextView.setText("Workload = " + String.format(Locale.getDefault(), "%6.2f", workload));
+        mTextView.setText("Workload = " + String.format(Locale.getDefault(), "%3d", workload));
     }
 
     @Override
